requirements.txt
simplegmail
pandas
google-cloud-bigquery
google-cloud-secret-manager
pandas-gbq
